Fjerner ubrukt kode
