+++ src/intTest/kotlin/no/nav/personbruker/dittnav/metrics/periodic/reporter/innboks/innboksQueriesTest.kt
@@ -2,15 +2,15 @@ package no.nav.personbruker.dittnav.metrics.periodic.reporter.innboks
 
 import kotlinx.coroutines.runBlocking
 import no.nav.personbruker.dittnav.metrics.periodic.reporter.common.database.H2Database
-import no.nav.personbruker.dittnav.metrics.periodic.reporter.done.DoneObjectMother
+import no.nav.personbruker.dittnav.metrics.periodic.reporter.common.database.util.countTotalNumberOfEvents
+import no.nav.personbruker.dittnav.metrics.periodic.reporter.common.database.util.countTotalNumberOfEventsByActiveStatus
+import no.nav.personbruker.dittnav.metrics.periodic.reporter.config.EventType
 import org.amshove.kluent.`should be equal to`
-import org.amshove.kluent.`should contain all`
-import org.amshove.kluent.`should be equal to`
-import org.amshove.kluent.`should not contain`
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Test
 
 class innboksQueriesTest {
+    
     private val database = H2Database()
 
     private val fodselsnummer1 = "12345"
@@ -19,9 +19,6 @@ class innboksQueriesTest {
     private val innboks1: Innboks
     private val innboks2: Innboks
     private val innboks3: Innboks
-
-    private val systembruker = "dummySystembruker"
-    private val eventId = "1"
 
     private val allInnboks: List<Innboks>
     private val allInnboksForAktor1: List<Innboks>
@@ -48,7 +45,7 @@ class innboksQueriesTest {
     }
 
     @AfterAll
-    fun cleanUp() {
+    fun tearDown() {
         runBlocking {
             database.dbQuery {
                 deleteAllInnboks()
@@ -56,114 +53,31 @@ class innboksQueriesTest {
         }
     }
 
+
     @Test
-    fun `finner alle Innboks`() {
+    fun `Skal telle det totale antall innboks-eventer`() {
         runBlocking {
             database.dbQuery {
-                val result = getAllInnboks()
-                result.size `should be equal to` allInnboks.size
-                result `should contain all` allInnboks
+                countTotalNumberOfEvents(EventType.INNBOKS)
             }
-        }
+        } `should be equal to` allInnboks.size.toLong()
     }
 
     @Test
-    fun `finner Innboks med id`() {
+    fun `Skal telle det totale antall aktive innboks-eventer`() {
         runBlocking {
             database.dbQuery {
-                val result = getInnboksById(innboks1.id!!)
-                result `should be equal to` innboks1
+                countTotalNumberOfEventsByActiveStatus(EventType.INNBOKS, true)
             }
-        }
+        } `should be equal to` allInnboks.size.toLong()
     }
 
     @Test
-    fun `setter aktiv flag`() {
-        val doneEvent = DoneObjectMother.giveMeDone(eventId, systembruker, fodselsnummer1)
+    fun `Skal telle det totale antall inaktive innboks-eventer`() {
         runBlocking {
             database.dbQuery {
-                setInnboksEventerAktivFlag(listOf(doneEvent), false)
-                var innboks = getInnboksByEventId(eventId)
-                innboks.aktiv `should be equal to` false
-
-                setInnboksEventerAktivFlag(listOf(doneEvent), true)
-                innboks = getInnboksByEventId(eventId)
-                innboks.aktiv `should be equal to` true
+                countTotalNumberOfEventsByActiveStatus(EventType.INNBOKS, false)
             }
-        }
-    }
-
-    @Test
-    fun `finner Innboks etter aktiv flag`() {
-        val doneEvent = DoneObjectMother.giveMeDone(innboks1.eventId, systembruker, fodselsnummer1)
-        runBlocking {
-            database.dbQuery {
-                setInnboksEventerAktivFlag(listOf(doneEvent), false)
-                val aktiveInnboks = getAllInnboksByAktiv(true)
-                val inaktivInnboks = getAllInnboksByAktiv(false)
-
-                aktiveInnboks.none { it.id == innboks1.id }
-                aktiveInnboks.size `should be equal to` allInnboks.size - 1
-                inaktivInnboks.single { it.id == innboks1.id }
-                inaktivInnboks.size `should be equal to` 1
-
-                setInnboksEventerAktivFlag(listOf(doneEvent), true)
-            }
-        }
-    }
-
-    @Test
-    fun `finner Innboks med fodselsnummer`() {
-        runBlocking {
-            database.dbQuery {
-                val result = getInnboksByFodselsnummer(fodselsnummer1)
-                result.size `should be equal to` allInnboksForAktor1.size
-                result `should contain all` allInnboksForAktor1
-                result `should not contain` innboks2
-            }
-        }
-    }
-
-    @Test
-    fun `finner Innboks med eventId`() {
-        runBlocking {
-            database.dbQuery {
-                val result = getInnboksByEventId(innboks1.eventId)
-                result `should be equal to` innboks1
-            }
-        }
-    }
-
-    @Test
-    fun `persister ikke entitet dersom rad med samme eventId og produsent finnes`() {
-        runBlocking {
-            database.dbQuery {
-                createInnboks(innboks1)
-                val numberOfEvents = getAllInnboks().size
-                createInnboks(innboks1)
-                getAllInnboks().size `should be equal to` numberOfEvents
-            }
-        }
-    }
-
-    @Test
-    fun `Skal skrive eventer i batch`() {
-        val innboks1 = InnboksObjectMother.giveMeAktivInnboks("i-1", "123")
-        val innboks2 = InnboksObjectMother.giveMeAktivInnboks("i-2", "123")
-
-        runBlocking {
-            database.dbQuery {
-                createInnboksEventer(listOf(innboks1, innboks2))
-            }
-
-            val innboks1FraDb = database.dbQuery { getInnboksByEventId(innboks1.eventId) }
-            val innboks2FraDb = database.dbQuery { getInnboksByEventId(innboks2.eventId) }
-
-            innboks1FraDb.eventId `should be equal to` innboks1.eventId
-            innboks2FraDb.eventId `should be equal to` innboks2.eventId
-
-            database.dbQuery { deleteInnboksWithEventId(innboks1.eventId) }
-            database.dbQuery { deleteInnboksWithEventId(innboks2.eventId) }
-        }
+        } `should be equal to` 0
     }
 }
